net/mlx5: VF group code adjustments
The patch removed redundant max rate configuration
of a VF vport after changing the group it belongs to.
The max rate is already set when creating the new scheduling
element for the vport.

In addition, adding a missing device rate limit capability check
to the VF group rate setting handler.

Issue: 1761999
Change-Id: I0cb508c96869564054844e9fd0a26c2cf9a01aa7
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>