RDMA/mlx5: Support TX port affinity for VF drivers in LAG mode
The mlx5 VF driver doesn't set QP tx port affinity in balance because it
doesn't know if the lag is active or not (the "lag_active" works only
for PF interfaces). In this case for VF interfaces only one lag is used
which brings performance issue.
So a lag_tx_port_affinity CAP bit is introduced in, when it is enabled
then driver always generate QP tx affinity, regardless of lag state.

Issue: 1695850
Change-Id: Ic982bf5406e270b9e01f454664365377b4fa0be8
Signed-off-by: Mark Zhang <markz@mellanox.com>