net/mlx5: Manage ldev with kref
We need to make sure we maintain ldev state when moving from/to
switchdev mode. Once a uplink representor was added it meant we free
the ldev and create a new one, use kref to manage the life cycle of the
ldev and when changing states (legcy <-> switchdev) take a ref so it
won't be freed.

Issue: 1706533
Change-Id: I11c2df1ae33503a19d194c6c28a15445f2d3df42
Signed-off-by: Mark Bloch <markb@mellanox.com>