Backports: Provide checking for pfmemalloc pages on specific kernels
On kernels having the pfmemalloc feature but lack of page_is_pfmemalloc(),
the mlx5 driver is unable to recognize the pfmemalloc pages. The
consequence is the pfmemalloc rx pages are recycled to the driver's
rx page cache instead of being returned to the kernel in the recycle
stage, which violates the design of pfmemalloc and may cause a series
of issues.

This patch fix the issue by providing an alternative way to inspect
pamemalloc pages for the kernels with pfmemalloc but without
page_is_pfmemalloc().

This patch also fix another issue that HAVE_PAGE_IS_PFMEMALLOC
is misused to identify the existence of numa_mem_id()
in mlx5e_page_is_reserved().

Issue: 1807817
Change-Id: Ibe062e8dcb788ffb63141c5eea5a29e88812e754
Signed-off-by: Fan Li <fanl@mellanox.com>