IB/ipoib: Fix possible deadlock over send queue
While napi callback and xmit functions are running in parallel it might
ends with closed send queue forever, the situation is as the following:

ipoib_send				ipoib_tx_poll(napi)
----------				------------------
closing tx queue			running, taking from the queue
					less than half of the queue-size
					not opening the queue
					end of napi loop
					less than budget
arming the CQ				....
CQ firing event (hw has all the wqe):
fails to schedule napi:
(already running)
					arming the CQ
					no more events from CQ
					already fired.

And now we are with closed queue, and nothing can't open it.

The fix is to make sure that when the xmit function re-arming the CQ in
closed queue the rescheduling will success.

Issue:1547200
Change-Id: I7bf33b26c5b2674403eef3ef095c44fb7b956a04
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>