net/mlx5: Fix use after free of command outbox
Currently command outbox is freed in completion handler and not in
cmd_exec only in case of callback command. Free the command outbox in
the completion handler also in case of non callback command to avoid use
after free in case completion event comes at the same time that
completion timeout occurs.

Issue: 1843020
Change-Id: I0712611dcab86ff851fc4a3c7abe27b002c708a5
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>