net/mlx5: Miniflow, Avoid free twice if the flow is freed by aging
The miniflow includes tc flows and ct flows. If tc flows are freed,
then all ct flows are freed as well. But if the tc flows are shared
by other ct flows, aging can delete the individual ct flows if there
is no traffic in 30s.

If the ct flows are freed by aging, we should not free them again
in miniflow_cleanup().

This patch add a flag. If it is true, we will not free ct flows
in miniflow_cleanup() again.

Issue: 1834196
Change-Id: Ic11841c9c37e86613bfcba3456ca26cb55eb6d1f
Signed-off-by: Chris Mi <chrism@mellanox.com>