memtrack: Fix memdup_user test for allocated address
The kernel's memdup_user() does not return NULL when it fails to allocate a 0
length buffer.  This caused memtrack's memdup_user macro to log the failed
address and later on report it as duplicate.  Replace the if(!addr) with
if(!ZERO_NOT_NULL_PTR(addr) in memtrack's memdup_user to solve.

Issue: 1856726
Change-Id: Ie668df01e7748f247a955784391578146b70466d
Signed-off-by: Yonatan Cohen <yonatanc@mellanox.com>