net/mlx5e: Fix out of memory when setting jumbo frames
Due to an improvement in Striding RQ feature, WQE stride size
(frag_size) is changed based on the MTU value. This change improves the
performance but also increases the dma memory request for large MTU
value.

As the result, when setting MTU value to 9000, PowerPC with
dynamic dma window disabled (ddw) is unable to satisfy the dma memory
request and returns error (iommu_alloc failed).

Solution is to reduce the dma memory request size on PowerPC with ddw
disabled by reducing the default rx rings from 1024 to 512. User
still can increase the rx rings via ethtool and see the issue but
there is nothing we can do for this.

Issue: 1602952
Change-Id: Ib9ace23498f7b06fd4fd0ab06f72ccf773bac5a1
Fixes: 73281b78a37a ("net/mlx5e: Derive Striding RQ size from MTU")
Signed-off-by: Huy Nguyen <huyn@mellanox.com>