IB/mlx4_core: Fix modify qp flow for XRC qps
Upstream kernel 5.0 commit 899444505473 ("IB/{hw,sw}: Remove
'uobject->context' dependency in object creation APIs") caused
__mlx4_ib_modify_qp() to treat XRC_TGT QP's as userspace QPs, rather than
kernel QPs (although are they created as kernel QPs).

This change broke mlnx_ofed 4.7 commit 454c85779702
("IB/mlx4: Assign user allocated DB and BF to QP"), which expects
pointer qp->uar to be non-null for userspace QPs in procedure
__mlx4_ib_modify_qp(). In this commit, qp->uar is only initialized
when creating userspace QPs. When creating kernel QPs, it remains NULL.
Therefore, when calling __mlx4_ib_modify_qp() for an XRC TGT qp
(which is treated here as a userspace QP), we see a NULL-reference Oops.

Since XRC_TGT QPs do not send or receive data, fix the problem
by using the ucontext uar index (passed to procedure __mlx4_ib_modify_qp)
instead of qp->uar->index when modifying XRC_TGT QPs. This has the
advantage of not exposing the kernel's uar index to the XRC_TGT qp.

NOTE: This patch is a *temporary fix*, which works for mlnx_ofed 4.7.
The basic problem is that the upstream kernel patch which removed the
uobject->context dependency when creating QPs makes an assumption that
no userspace app causes creation of a kernel-space QP, and then calls
modify_qp on that QP. For XRC TGT QPs, this assumption is false.

I will be opening a discussion thread in the upstream kernel regarding
how to deal with XRC TGT QPs (which are created as kernel QPs, but in
the upstream kernel are modified as userspace QPs since kernel 5.0.

Furthermore, when we eliminate the experimental verbs from mlnx_ofed,
this patch should also be removed, along with the patch assigning
user allocated DB/BFs.

issue: 1856817
Change-Id: I6b107b2027f8591f3012428149d18cbbf91818cd
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>