BACKPORTS: fix user-mad.c backport
For the kernels which do not have HAVE_IDA_ALLOC_MAX defined,
the backport code was missing a "return -1" line.

As a result, the devnum for the mad devices remained zero,
and  infiniband_mad/umad1 failed to come up, because the
port->devnum remained zero, so the device name used
for port 2 was umad0, instead of umad1.

issue:1890642
Change-Id: Ifd2a5c5ae731ee0e45290cbae401785f460a5892
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>