IB/mlx5: Test write combining support
Change reporting of BlueFlame support to be dependent on write-combining
support.

Add a test in mlx5_ib initialization process to test whether
write-combining is supported on the machine.
The test will run as part of the enable_driver callback to ensure that
the test runs after the device is setup and can create and modify the
QP needed, but runs before the device is exposed to the users.

The test opens UD QP and posts NOP WQEs. Inside the post-send we modify
the WQE written to the BlueFlame from the WQE in memory, requesting CQE
only on the BlueFlame WQE.
By checking whether we received a completion on one of these WQEs we can
know if BlueFlame succeeded and whether write-combining is supported.

Issue: 1846808
Change-Id: I69cdb4d9897a9677f46f70832aaa8b2ac779bed0
Signed-off-by: Michael Guralnik <michaelgur@mellanox.com>