net/mlx5: E-Switch, Add bond_metadata and its shadow entries
Adding bond_metadata and its shadow entries to represent a lag device
and its slaves VF representors. Bond_metadata structure includes a
unique metadata shared by slaves VF respresentors, and a list of slaves
representors shadow entries.

On enslaving event, create a bond_metadata structure representing
the upper lag device of this slave representor if it has not been
created yet. Create and add entry for the slave representor to the
slaves list. Elect the first slave to be the "primary" representor
so that other slave representors will share its ingress acl table.

On unslaving event, free the shadow entry of the slave representor.
If this slave representor is the primary one, elect the new primary
slave in the slave list as replacement.
On the last unslave event, free the bond_metadata structure and its
resources.

Introduce APIs to create and remove bond_metadata and its resources,
enslave and unslave VF representor shadow entries.

Issue: 1334821
Change-Id: Ibc4020fe775f6a9f2b09fa7bc848207068fe05a6
Signed-off-by: Vu Pham <vuhuong@mellanox.com>