net/mlx5e: miniflow, Hold dep_lock when freeing flow
When miniflow aging is working, it holds the dep_lock. But
in mlx5e_flow_put, we don't hold the same lock. So the list
may be corrupted by mlx5e_flow_put.

This commit also adds a new function to deal with the case
that dep_lock is already held.

Issue: 1903601
Change-Id: I2eaa19763b869507b3846cc04f78aedb626d25d2
Signed-off-by: Chris Mi <chrism@mellanox.com>