Backports: Avoid use of local variable in deferred code
Use a static local variable whose allocation is not freed to pass to
eswitch mode change code which runs in the context of a task where you
don't know whether it will be used after the caller function's scope was
exited.
Using static is safe is this code is very slow to happen as it is used
for mode change.

Issue: 1910930
Change-Id: If4d04d1ddbd9a7faae5dc64d59685600d8f9f727
Signed-off-by: Eli Cohen <eli@mellanox.com>