BACKPORTS: Disable pernet and namespace in core device for old kernels
In older kernels (for example, RHEL6.10 and SLES11 SP4) if
pernet operations are enabled in device.c, we get an Oop
at driver startup.

Fix this by verifying that the pernet_operations struct
has an id member, and that struct class has a ns_type member.

If either of these conditions is not true, then disable
the pernet and namespace features for device.c

issue: 1915162
Change-Id: I95468f0422e8038ba8a5e4b336fca56be273f0e5
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>